requirements.txt
django-filter==21.1
djangorestframework==3.14.0
djoser==2.1.0
# python-decouple==3.5
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.3
drf-extra-fields==3.5.0
filetype==1.2.0
# pandas
numpy==1.25.0
pandas==2.0.2
python-dateutil==2.8.2
tzdata==2023.3

django-cors-headers==4.1.0
